pyproject.toml
[tool.poetry]
name = "jota-python-utils"
version = "0.1.0"
description = "Utils library for JOTA backend apps."
authors = ["Tecnologia JOTA <tecnologia@jota.info>"]

[tool.poetry.dependencies]
python = ">= 3.7"
Django = "3.2.14"
django-phonenumbers = "^1.0.1"
django-phonenumber-field = "^6.3.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
